requirements.txt
# Core FastAPI dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0

# CORS middleware (built into FastAPI)
# fastapi-cors==0.0.6

# Data processing
numpy==1.24.3
pandas==2.0.3

# Machine Learning
torch>=2.8.0
torchvision>=0.23.0
transformers>=4.57.0
open-clip-torch==2.20.0
scikit-learn>=1.3.0

# Vector Database
pymilvus==2.3.4

# Image processing
Pillow==10.0.1

# HTTP requests
requests==2.31.0

# Progress bars
tqdm==4.66.1

# Concurrent processing (built-in for Python 3.9+)
# concurrent-futures==3.1.1

# Environment variables
python-dotenv==1.0.0

# Development dependencies
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.9.1
flake8==6.1.0

# Compatibility fixes
marshmallow==3.20.2
